refactor (se borran los espacios en blancos)

+++ Trabajo Entregable 3/src/main/java/com/example/trabajoentregable3/dto/DTOReport.java
@@ -3,15 +3,15 @@ package com.example.trabajoentregable3.dto;
 public class DTOReport {
 	private String career_name;
 	private Integer year;
-	private Double enrolled_qty;
-	private Double graduated_qty;
+	private Integer enrolled_qty;
+	private Integer graduated_qty;
 	
 	public DTOReport(String career_name, Integer year, Double cant_inscriptos, Double cant_egresos) {
 		super();
 		this.career_name = career_name;
 		this.year = year;
-		this.enrolled_qty = cant_inscriptos;	
-		this.graduated_qty = cant_egresos;
+		this.enrolled_qty = cant_inscriptos.intValue();
+		this.graduated_qty = cant_egresos.intValue();
 	}	
 
 	public String getCareer_name() {
@@ -30,26 +30,26 @@ public class DTOReport {
 		this.year = year;
 	}
 
-	public Double getEnrolled_qty() {
+	public Integer getEnrolled_qty() {
 		return enrolled_qty;
 	}
 
-	public void setEnrolled_qty(Double enrolled_qty) {
+	public void setEnrolled_qty(Integer enrolled_qty) {
 		this.enrolled_qty = enrolled_qty;
 	}
 
-	public Double getGraduated_qty() {
+	public Integer getGraduated_qty() {
 		return graduated_qty;
 	}
 
-	public void setGraduated_qty(Double graduated_qty) {
+	public void setGraduated_qty(Integer graduated_qty) {
 		this.graduated_qty = graduated_qty;
 	}
 
 	@Override
 	public String toString() {
-		return "Carrera: " + career_name + ", Año: " + year + ", Cantidad Inscriptos: " + enrolled_qty.intValue()
-				+ ", Cantidad Egresos: " + graduated_qty.intValue() + " \n";
+		return "Carrera: " + career_name + ", Año: " + year + ", Cantidad Inscriptos: " + enrolled_qty
+				+ ", Cantidad Egresos: " + graduated_qty + " \n";
 	}
 	
 }
